Improved support for binding a mouse buttons.
The keydown/up hooks now receive the target view in addition to the key. The default keymap functionality chooses which parameters to pass based on the arity of the mapped function:

0: (fn)
1: (fn view)
2: (fn key view)

These things combined let the view-focus hook be added as:

(define-key! default-keymap (kbd "Mouse1") view-focus)

However, unfortunately there is no way to override the swallow-next-key
call that happens when a key is matched by the keymap. Forcing a return
value from each function called seems unnecessary. Probably need an
opt-out way to avoid it, because it is sensible default behavior.